+++ backend/canisters/community/impl/src/model/members.rs
@@ -1,11 +1,10 @@
 use crate::model::members::stable_memory::MembersStableStorage;
-use crate::model::members_map::{HeapMembersMap, MembersMap};
 use crate::model::user_groups::{UserGroup, UserGroups};
 use candid::Principal;
 use group_community_common::{Member, MemberUpdate, Members};
 use rand::RngCore;
 use serde::{Deserialize, Serialize};
-use std::collections::{BTreeMap, BTreeSet, VecDeque};
+use std::collections::{BTreeMap, BTreeSet};
 use types::{
     is_default, ChannelId, CommunityMember, CommunityPermissions, CommunityRole, TimestampMillis, Timestamped, UserId,
     UserType, Version,
@@ -19,8 +18,8 @@ const MAX_MEMBERS_PER_COMMUNITY: u32 = 100_000;
 
 #[derive(Serialize, Deserialize)]
 pub struct CommunityMembers {
-    members: HeapMembersMap,
-    stable_memory_members_map: MembersStableStorage,
+    #[serde(alias = "stable_memory_members_map")]
+    members_map: MembersStableStorage,
     member_channel_links: BTreeSet<(UserId, ChannelId)>,
     member_channel_links_removed: BTreeMap<(UserId, ChannelId), TimestampMillis>,
     user_groups: UserGroups,
@@ -36,8 +35,6 @@ pub struct CommunityMembers {
     members_with_display_names: BTreeSet<UserId>,
     members_with_referrals: BTreeSet<UserId>,
     updates: BTreeSet<(TimestampMillis, UserId, MemberUpdate)>,
-    migrate_to_stable_memory_queue: VecDeque<UserId>,
-    migration_to_stable_memory_complete: bool,
 }
 
 impl CommunityMembers {
@@ -62,8 +59,7 @@ impl CommunityMembers {
         };
 
         CommunityMembers {
-            members: HeapMembersMap::new(member.clone()),
-            stable_memory_members_map: MembersStableStorage::new(member),
+            members_map: MembersStableStorage::new(member),
             member_channel_links: public_channels.into_iter().map(|c| (creator_user_id, c)).collect(),
             member_channel_links_removed: BTreeMap::new(),
             user_groups: UserGroups::default(),
@@ -82,8 +78,6 @@ impl CommunityMembers {
             members_with_display_names: BTreeSet::new(),
             members_with_referrals: BTreeSet::new(),
             updates: BTreeSet::new(),
-            migrate_to_stable_memory_queue: VecDeque::default(),
-            migration_to_stable_memory_complete: true,
         }
     }
 
@@ -116,12 +110,12 @@ impl CommunityMembers {
                 referrals: BTreeSet::new(),
                 lapsed: Timestamped::default(),
             };
-            self.insert_internal(member.clone());
+            self.members_map.insert(member.clone());
             self.add_user_id(principal, user_id);
 
             if let Some(referrer) = referred_by {
                 if matches!(
-                    self.update_member_internal(&referrer, |m| {
+                    self.update_member(&referrer, |m| {
                         m.referrals.insert(user_id);
                         true
                     }),
@@ -147,7 +141,7 @@ impl CommunityMembers {
 
     pub fn remove_by_principal(&mut self, principal: &Principal, now: TimestampMillis) -> Option<CommunityMemberInternal> {
         if let Some(user_id) = self.principal_to_user_id_map.remove(principal) {
-            if let Some(member) = self.remove_internal(&user_id) {
+            if let Some(member) = self.members_map.remove(&user_id) {
                 match member.role {
                     CommunityRole::Owner => self.owners.remove(&user_id),
                     CommunityRole::Admin => self.admins.remove(&user_id),
@@ -170,7 +164,7 @@ impl CommunityMembers {
                 }
                 if let Some(referrer) = member.referred_by {
                     let mut remove_from_members_with_referrals = false;
-                    self.update_member_internal(&referrer, |m| {
+                    self.update_member(&referrer, |m| {
                         m.referrals.remove(&user_id);
                         if m.referrals.is_empty() {
                             remove_from_members_with_referrals = true;
@@ -224,7 +218,7 @@ impl CommunityMembers {
             None => return ChangeRoleResult::UserNotInCommunity,
         }
 
-        let mut member = match self.get_internal(&target_user_id) {
+        let mut member = match self.members_map.get(&target_user_id) {
             Some(p) => p,
             None => return ChangeRoleResult::TargetUserNotInCommunity,
         };
@@ -269,7 +263,7 @@ impl CommunityMembers {
             _ => false,
         };
 
-        self.insert_internal(member);
+        self.members_map.insert(member);
 
         ChangeRoleResult::Success(ChangeRoleSuccessResult {
             caller_id: user_id,
@@ -278,7 +272,7 @@ impl CommunityMembers {
     }
 
     pub fn set_suspended(&mut self, user_id: UserId, suspended: bool, now: TimestampMillis) -> Option<bool> {
-        let result = self.update_member_internal(&user_id, |member| {
+        let result = self.update_member(&user_id, |member| {
             if member.suspended.value != suspended {
                 member.suspended = Timestamped::new(suspended, now);
                 true
@@ -371,7 +365,7 @@ impl CommunityMembers {
     }
 
     pub fn mark_rules_accepted(&mut self, user_id: &UserId, version: Version, now: TimestampMillis) {
-        self.update_member_internal(user_id, |member| member.accept_rules(version, now));
+        self.update_member(user_id, |member| member.accept_rules(version, now));
     }
 
     pub fn block(&mut self, user_id: UserId) -> bool {
@@ -410,11 +404,11 @@ impl CommunityMembers {
 
         let user_id = self.principal_to_user_id_map.get(&user_id_or_principal).unwrap_or(&user_id);
 
-        self.get_internal(user_id)
+        self.members_map.get(user_id)
     }
 
     pub fn get_by_user_id(&self, user_id: &UserId) -> Option<CommunityMemberInternal> {
-        self.get_internal(user_id)
+        self.members_map.get(user_id)
     }
 
     // Note this lookup is O(n)
@@ -486,7 +480,7 @@ impl CommunityMembers {
     pub fn set_display_name(&mut self, user_id: UserId, display_name: Option<String>, now: TimestampMillis) {
         let display_name_is_some = display_name.is_some();
         if matches!(
-            self.update_member_internal(&user_id, |m| {
+            self.update_member(&user_id, |m| {
                 m.display_name = Timestamped::new(display_name, now);
                 true
             }),
@@ -503,7 +497,7 @@ impl CommunityMembers {
 
     pub fn update_lapsed(&mut self, user_id: UserId, lapsed: bool, now: TimestampMillis) {
         if matches!(
-            self.update_member_internal(&user_id, |m| {
+            self.update_member(&user_id, |m| {
                 if lapsed {
                     // Owners can't lapse
                     !m.is_owner() && m.set_lapsed(true, now)
@@ -529,10 +523,7 @@ impl CommunityMembers {
 
     pub fn unlapse_all(&mut self, now: TimestampMillis) {
         for user_id in std::mem::take(&mut self.lapsed) {
-            if matches!(
-                self.update_member_internal(&user_id, |m| m.set_lapsed(false, now)),
-                Some(true)
-            ) {
+            if matches!(self.update_member(&user_id, |m| m.set_lapsed(false, now)), Some(true)) {
                 self.updates.insert((now, user_id, MemberUpdate::Unlapsed));
             }
         }
@@ -556,40 +547,22 @@ impl CommunityMembers {
         .unwrap()
     }
 
-    fn get_internal(&self, user_id: &UserId) -> Option<CommunityMemberInternal> {
-        if self.migration_to_stable_memory_complete {
-            self.stable_memory_members_map.get(user_id)
-        } else {
-            self.members.get(user_id)
-        }
-    }
-
-    fn insert_internal(&mut self, member: CommunityMemberInternal) {
-        self.members.insert(member.clone());
-        self.stable_memory_members_map.insert(member);
-    }
-
-    fn remove_internal(&mut self, user_id: &UserId) -> Option<CommunityMemberInternal> {
-        self.stable_memory_members_map.remove(user_id);
-        self.members.remove(user_id)
-    }
-
-    fn update_member_internal<F: FnOnce(&mut CommunityMemberInternal) -> bool>(
+    fn update_member<F: FnOnce(&mut CommunityMemberInternal) -> bool>(
         &mut self,
         user_id: &UserId,
         update_fn: F,
     ) -> Option<bool> {
-        let mut member = self.get_internal(user_id)?;
+        let mut member = self.members_map.get(user_id)?;
 
         let updated = update_fn(&mut member);
         if updated {
-            self.insert_internal(member);
+            self.members_map.insert(member);
         }
         Some(updated)
     }
 
     #[cfg(test)]
-    fn check_invariants(&self, stable_memory_map: bool) {
+    fn check_invariants(&self) {
         let mut member_ids = BTreeSet::new();
         let mut owners = BTreeSet::new();
         let mut admins = BTreeSet::new();
@@ -598,13 +571,7 @@ impl CommunityMembers {
         let mut members_with_display_names = BTreeSet::new();
         let mut members_with_referrals = BTreeSet::new();
 
-        let all_members = if stable_memory_map {
-            self.stable_memory_members_map.all_members()
-        } else {
-            self.members.all_members()
-        };
-
-        for member in all_members {
+        for member in self.members_map.all_members() {
             member_ids.insert(member.user_id);
 
             match member.role {
@@ -663,25 +630,25 @@ impl Members for CommunityMembers {
 
 #[derive(Serialize, Deserialize, Clone)]
 pub struct CommunityMemberInternal {
-    #[serde(rename = "u", alias = "user_id")]
+    #[serde(rename = "u")]
     pub user_id: UserId,
-    #[serde(rename = "d", alias = "date_added")]
+    #[serde(rename = "d")]
     pub date_added: TimestampMillis,
-    #[serde(rename = "r", alias = "role", default, skip_serializing_if = "is_default")]
+    #[serde(rename = "r", default, skip_serializing_if = "is_default")]
     role: CommunityRole,
-    #[serde(rename = "ra", alias = "rules_accepted", skip_serializing_if = "Option::is_none")]
+    #[serde(rename = "ra", skip_serializing_if = "Option::is_none")]
     pub rules_accepted: Option<Timestamped<Version>>,
-    #[serde(rename = "ut", alias = "user_type", default, skip_serializing_if = "is_default")]
+    #[serde(rename = "ut", default, skip_serializing_if = "is_default")]
     pub user_type: UserType,
-    #[serde(rename = "dn", alias = "display_name", default, skip_serializing_if = "is_default")]
+    #[serde(rename = "dn", default, skip_serializing_if = "is_default")]
     display_name: Timestamped<Option<String>>,
-    #[serde(rename = "rb", alias = "referred_by", skip_serializing_if = "Option::is_none")]
+    #[serde(rename = "rb", skip_serializing_if = "Option::is_none")]
     pub referred_by: Option<UserId>,
-    #[serde(rename = "rf", alias = "referrals", default, skip_serializing_if = "BTreeSet::is_empty")]
+    #[serde(rename = "rf", default, skip_serializing_if = "BTreeSet::is_empty")]
     referrals: BTreeSet<UserId>,
-    #[serde(rename = "l", alias = "lapsed", default, skip_serializing_if = "is_default")]
+    #[serde(rename = "l", default, skip_serializing_if = "is_default")]
     lapsed: Timestamped<bool>,
-    #[serde(rename = "s", alias = "suspended", default, skip_serializing_if = "is_default")]
+    #[serde(rename = "s", default, skip_serializing_if = "is_default")]
     suspended: Timestamped<bool>,
 }