+++ backend/canisters/community/impl/src/model/mod.rs
@@ -3,6 +3,5 @@ pub mod events;
 pub mod groups_being_imported;
 pub mod invited_users;
 pub mod members;
-pub mod members_map;
 pub mod user_event_batch;
 pub mod user_groups;